refactor - refactoring login api file structure
