+++ types/apiTypes/product.d.ts
@@ -44,11 +44,5 @@ declare namespace res {
       views: number;
     };
   };
-  type ProductDeleteError = {
-    message: string;
-    status: number;
-    errors: string[];
-    code: string;
-  };
   type ProductStatus = 'soldout' | 'sale' | 'reserve';
 }
